Last Revisions, Code Cleanup, Commenting
Made 3 requested fixes: The asterisked phrase, getting rid of a colon, and added the glow hover effect from get info to other buttons.

I also brought in the normalize css file as a base reset class file.

I went through and commented up the code by way of documenting the JS functionality, and the same for the animation css classes.

Lastly, I've gone through and reorganized the CSS, so that the class definitioins are in a somewhat predictable manner.